+++ apps/site/app/routes/$.tsx
@@ -4,4 +4,7 @@ export { default, loader } from "./index";
 
 export const meta: MetaFunction = ({ data }) => ({
   title: data?.title ? `${data.title} - Superflare` : "Superflare",
+  "twitter:title": data?.title ? `${data.title} - Superflare` : "Superflare",
+  description: data?.description,
+  "twitter:description": data?.description,
 });