+++ packages/superflare/cli/wrangler.ts
@@ -1,29 +1,59 @@
 import { spawn } from "node:child_process";
 
+export type WranglerCommandResponse = {
+  code: number;
+  stdout: string;
+  stderr: string;
+};
+
+/**
+ * Run a wrangler command. It would be great to replace this with a real exported API instead
+ * of spawning a child process every time.
+ */
+export async function runWranglerCommand(
+  command: string[]
+): Promise<WranglerCommandResponse> {
+  let stdout = "";
+  let stderr = "";
+
+  const child = spawn("npx", ["wrangler@latest", ...command], {
+    shell: true,
+    env: {
+      ...process.env,
+      // TODO: Remove when D1 is stable.
+      NO_D1_WARNING: "true",
+    },
+  });
+
+  child.stderr.on("data", (data) => {
+    stderr += data;
+  });
+  child.stdout.on("data", (data) => {
+    stdout += data;
+  });
+
+  return new Promise((resolve, reject) => {
+    child.on("close", (code) => {
+      if (code === 0) {
+        resolve({ code, stdout, stderr });
+        return;
+      }
+
+      reject({ code, stdout, stderr });
+    });
+  });
+}
+
 /**
  * Spawn a child process to execute: npx wrangler d1 migrations apply DB --local
  */
 export async function wranglerMigrate() {
-  return new Promise<void>((resolve, reject) => {
-    spawn(
-      "npx",
-      ["wrangler", "d1", "migrations", "apply", "DB", "--local", "-j"],
-      {
-        env: {
-          ...process.env,
-          CI: "true",
-        },
-      }
-    ).on("close", (code) => {
-      if (code === 0) {
-        resolve();
-      } else {
-        reject(
-          new Error(
-            `npx wrangler d1 migrations apply DB --local exited with code ${code}`
-          )
-        );
-      }
-    });
-  });
+  return await runWranglerCommand([
+    "d1",
+    "migrations",
+    "apply",
+    "DB",
+    "--local",
+    "-j",
+  ]);
 }