+++ templates/remix/db/migrations/0000_create_users.ts
@@ -1,0 +1,10 @@
+import { Schema } from "superflare";
+
+export default function () {
+  return Schema.create("users", (table) => {
+    table.increments("id");
+    table.string("email").unique();
+    table.string("password");
+    table.timestamps();
+  });
+}