+++ templates/remix/db/seed.ts
@@ -1,0 +1,5 @@
+import { seed } from "superflare";
+
+export default seed(async () => {
+  //
+});